third_party/blink/web_tests/virtual/focusless-spat-nav/fast/spatial-navigation/focusless/snav-focusless-form-state.html

<!DOCTYPE html>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/page/spatial_navigation.mojom.js"></script>
<script src="../../../../../fast/spatial-navigation/resources/mock-snav-service.js"></script>
<script src="../../../../../fast/spatial-navigation/resources/snav-testharness.js"></script>

<form>
<input type="text" id="first" autofocus></input>
<input type="text" id="second"></input>
<input type="text" id="third"></input>
</form>

<script>
  async_test((t) => {
    let first = document.getElementById("first");
    let second = document.getElementById("second");
    let third = document.getElementById("third");

    let initialState = snavCallback();

    async function runTest() {
      assert_equals(document.activeElement,
                    first,
                    "'first' should start focused.");
      assert_equals(window.internals.interestedElement,
                    first,
                    "'first' should start interested.");

      await initialState;
      assert_true(mockSnavService.canExitFocus,
                  "Should be able to exit focus.");
      assert_true(mockSnavService.canSelectElement,
                  "Should be able to select element.");
      assert_true(mockSnavService.hasNextFormElement,
                  "Should be able to move to next form element.");

      mockSnavService.callback = notCalled;
      eventSender.keyDown('Tab');
      assert_equals(document.activeElement,
                    second,
                    "'second' should be focused.");

      await snav.rAF();

      eventSender.keyDown('Tab');
      assert_equals(document.activeElement,
                    third,
                    "'third' should be focused.");

      await snavCallback();
      assert_true(mockSnavService.canExitFocus,
                  "Should be able to exit focus.");
      assert_true(mockSnavService.canSelectElement,
                  "Should be able to select element.");
      assert_false(mockSnavService.hasNextFormElement,
                   "Should be able to move to next form element.");

      eventSender.keyDown('Escape');

      await snavCallback();
      assert_false(mockSnavService.canExitFocus,
                   "Should be able to exit focus.");
      assert_true(mockSnavService.canSelectElement,
                  "Should be able to select element.");
      assert_false(mockSnavService.hasNextFormElement,
                   "Should be able to move to next form element.");

      t.done();
    };

    t.step_timeout(() => {
      runTest();
    }, 0);

    function notCalled() {
      t.step_timeout(() => {
        assert_false(true, "Should not be called");
      }, 0);
    };

  }, 'Spat Nav state updates correctly for form elements.');
</script>
